+++ app/scripts/configs/tracks-info-by-type.js
@@ -2,6 +2,11 @@ import { TRACKS_INFO } from '.';
 
 export const TRACKS_INFO_BY_TYPE = TRACKS_INFO.reduce((tracksByType, track) => {
   tracksByType[track.type] = track;
+  if (track.aliases) {
+    for (const alias of track.aliases) {
+      tracksByType[alias] = track;
+    }
+  }
   return tracksByType;
 }, {});